requirements.txt
keyboard==0.13.5
numpy==2.2.5
Pillow==11.2.1
protobuf==5.29.4
pywin32==310
PyYAML==6.0.2
soundcard==0.4.4
soundfile==0.13.1
webrtcvad==2.0.10
webrtcvad-wheels==2.0.14
setuptools>=65.0.0
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
pydantic==2.11.5
tqdm==4.67.1
typing-extensions==4.13.2
proto-plus==1.26.1
Cerberus==1.3.5
